Fix variadic arguments on arm64 darwin ABI
Normal arguments that spill on the stack are packed, but not variadic
arguments. This is handled correctly for their placement already, but
code generated on the callee side with va_list expects word-size
sign-extension, so we need to fill the entire word.

(cherry picked from commit e42520ca5be190c9be207355b8afc9e4401bfcdd)